+++ srcs/ogv1.c
@@ -61,9 +61,9 @@ int main(int argc, char* argv[])
 
 	mlx.handle = mlx_init();
 	mlx.window = mlx_new_window(mlx.handle, WINDOW_WIDTH, WINDOW_HEIGHT, "cuv3d");
-	mlx.player.x = 250;
-	mlx.player.y = 250;
-	mlx.player.angle = 63.7;
+	mlx.player.x = 308;
+	mlx.player.y = 312;
+	mlx.player.angle = 90;//63.7
 	mlx.player.pdx = cos_wrap(mlx.player.angle);
 	mlx.player.pdy = -sin_wrap(mlx.player.angle);
 	mlx.player.side = 10;
